Fix reset when the storage quota is low
FossilOrigin-Name: 131b5910610ba09103ce93f5b823266175ad73a0e6f8b0fa17b81e3e92e08db1

+++ src/include/lib/Garradin/Utils.php
@@ -667,7 +667,7 @@ class Utils
 
             if (is_dir($path . DIRECTORY_SEPARATOR . $file))
             {
-                if (!self::deleteRecursive($path . DIRECTORY_SEPARATOR . $file))
+                if (!self::deleteRecursive($path . DIRECTORY_SEPARATOR . $file, true))
                     return false;
             }
             else
@@ -677,7 +677,10 @@ class Utils
         }
 
         $dir->close();
-        rmdir($path);
+
+        if ($delete_self) {
+            rmdir($path);
+        }
 
         return true;
     }
